# swagger.yaml
openapi: 3.0.3
info:
  title: SW2026 E-commerce API
  description: In-memory REST API for user authentication and checkout.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /checkout:
    post:
      tags:
        - Checkout
      summary: Checkout with cash or credit card
      description: >
        Requires a valid JWT. Cash payments receive a 10% discount.
        Credit card payments are charged at full price.
      operationId: checkout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
            examples:
              cash:
                summary: Cash checkout
                value:
                  paymentMethod: cash
                  items:
                    - productId: 1
                      quantity: 1
                    - productId: 3
                      quantity: 2
              creditCard:
                summary: Credit card checkout
                value:
                  paymentMethod: credit_card
                  items:
                    - productId: 2
                      quantity: 1
      responses:
        '200':
          description: Checkout completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '400':
          description: Invalid checkout payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
...
    CheckoutItemInput:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    CheckoutRequest:
      type: object
      required:
        - paymentMethod
        - items
      properties:
        paymentMethod:
          type: string
          enum:
            - cash
            - credit_card
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CheckoutItemInput'
    CheckoutLineItem:
      type: object
      properties:
        productId:
          type: integer
        name:
          type: string
        unitPrice:
          type: number
        quantity:
          type: integer
        lineTotal:
          type: number
    CheckoutResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        items:
          type: array
          items:
            $ref: '#/components/schemas/CheckoutLineItem'
        paymentMethod:
          type: string
          enum:
            - cash
            - credit_card
        subtotal:
          type: number
        discount:
          type: number
        discountRate:
          type: number
        total:
          type: number
    ErrorResponse:
      type: object
      properties:
        error:
          type: string